fix(channels): apply SSRF egress vetting to inbound IMAP connects
Extract the routability predicate and operator allowlist from safe_http
into a transport-agnostic egress module, leaving safe_http as the reqwest
adapter over it. Add resolve_and_validate for raw-TCP connectors.

IMAP open_session now resolves and vets the admin-supplied host, then
connects to a validated address while keeping the hostname for TLS, so a
channel can no longer dial a loopback, private, or cloud-metadata address.
Self-host reaches an internal mail server via NOSDESK_OUTBOUND_ALLOWED_HOSTS.
